+++ PersonalInquiryManager/Inquiries/INQMyInquiriesTableViewController.m
@@ -94,7 +94,7 @@ typedef NS_ENUM(NSInteger, inquiries) {
         NSUInteger cntAfter = [[self.fetchedResultsController fetchedObjects] count];
         
         if (cntBefore!=cntAfter) {
-            DLog(@"Inquiries: %d -> %d", cntBefore, cntAfter);
+            DLog(@"Inquiries: %lu -> %lu", (unsigned long)cntBefore, (unsigned long)cntAfter);
             
             [self.tableView reloadData];
         }
@@ -124,7 +124,7 @@ typedef NS_ENUM(NSInteger, inquiries) {
         NSUInteger cntAfter = [[self.fetchedResultsController fetchedObjects] count];
         
         // if (cntBefore!=cntAfter) {
-        DLog(@"Inquiries: %d -> %d", cntBefore, cntAfter);
+        DLog(@"Inquiries: %lu -> %lu", (unsigned long)cntBefore, (unsigned long)cntAfter);
         
         [self.tableView reloadData];
         // }
